// src/generated/models/CreateAitDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AIT  doc
 * AIT documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateAitDto
 */
export interface CreateAitDto {
    /**
     * Vehicle license plate
     * @type {string}
     * @memberof CreateAitDto
     */
    placaVeiculo: string;
    /**
     * Date and time of the infraction
     * @type {Date}
     * @memberof CreateAitDto
     */
    dataInfracao: Date;
    /**
     * Description of the infraction
     * @type {string}
     * @memberof CreateAitDto
     */
    descricao: string;
    /**
     * Fine amount in decimal
     * @type {number}
     * @memberof CreateAitDto
     */
    valorMulta: number;
}

/**
 * Check if a given object implements the CreateAitDto interface.
 */
export function instanceOfCreateAitDto(value: object): value is CreateAitDto {
    if (!('placaVeiculo' in value) || value['placaVeiculo'] === undefined) return false;
    if (!('dataInfracao' in value) || value['dataInfracao'] === undefined) return false;
    if (!('descricao' in value) || value['descricao'] === undefined) return false;
    if (!('valorMulta' in value) || value['valorMulta'] === undefined) return false;
    return true;
}

export function CreateAitDtoFromJSON(json: any): CreateAitDto {
    return CreateAitDtoFromJSONTyped(json, false);
}

export function CreateAitDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateAitDto {
    if (json == null) {
        return json;
    }
    return {
        
        'placaVeiculo': json['placa_veiculo'],
        'dataInfracao': (new Date(json['data_infracao'])),
        'descricao': json['descricao'],
        'valorMulta': json['valor_multa'],
    };
}

export function CreateAitDtoToJSON(json: any): CreateAitDto {
    return CreateAitDtoToJSONTyped(json, false);
}

export function CreateAitDtoToJSONTyped(value?: CreateAitDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'placa_veiculo': value['placaVeiculo'],
        'data_infracao': ((value['dataInfracao']).toISOString()),
        'descricao': value['descricao'],
        'valor_multa': value['valorMulta'],
    };
}